HomeScreen: Make a picture as background of menu window
Currently, the backgrund of menu window is a white transparent
rectangle. If some windows are in application layer, we could
see this app window from menu window. So the menu window
looks mess and seems something wrong in layer management.

This patch set a picture as background of menu window,
So user can't see app window from menu window.

Change-Id: I8ae049344ac2d87e0a27bb7b829d70b9b04faddf
Signed-off-by: Xiong Zhang <xiong.y.zhang@intel.com>