Fix FTBFS for x86_64
In all cases, the build failure was caused by a 64bit
pointer being casted to an int and then int being used
in debug log formatting (%x).

Better solution is to use %p and cast pointers to
void *.

Additionally, introcude a replacement for reinterpret_cast<>()
to handle int/ptr conversion.

Change-Id: I66e5cc2b3645210a675e27f8f82f7b12a81ac0e1
Signed-off-by: Mikko Ylinen <mikko.ylinen@intel.com>