bug fix: TC-115 (same cause TC-89)
TC-115 - Click app has no response after switching app for many times.
(same cause TC-89)
EFL has managed the grab of a pointer per drawing parts.
However, if a pointer is moved to another surface while the touch panel
was pushed and drawing parts have held the glab, the event which
detached the touch panel forever will not come to EFL.
Although a means to avoid this is prepared for EFL, since the
processing was not mounted in HomeScreen, while the parts pushed first
have held the glab, even if it pushes other parts, touch operation
loses its effect.
When repair moved to surface of an exception, with a touch panel
pushed, it corresponded by releasing the glab of EFL.

Change-Id: I2afb85eb14f9814b533f67c658755ad63a63ad24
Bug-Tizen: TC-115 TC-89
Signed-off-by: Ryota Okubi <ryota.okubi@mail.toyota-td.jp>