bug fix: TC-127(TIVI-3101)
TC-127 - No notification pop up while incoming call comes on.
It is a phenomenon in which pop-up of OnScreen is not displayed by
notification.Since OnScreen will generate and show Surface if notification
comes, but SystemController does not assume that one application makes
two or more Surfaces.
Repair allowed the application to make two or more surfaces,
when control as which an application gives a unique title to each surface,
and an application shows/hides each surface was performed.
In addition, the flag(multisurface) which allows the attribute of the
application defined by app_attr.conf generation of two or more Surfaces
is added, and only an application is allowed this when this flag is true.

Change-Id: Ib2646fbf84f3594b791f9b2cd9e9d5d19eac4b7d
Bug-Tizen: TC-127
Signed-off-by: Ryota Okubi <ryota.okubi@mail.toyota-td.jp>