EGLWLMockNavigation is not displayed.
Although EGLWLMockNavigation is a client of GENIVI-LM and surface is used
using API of GENIVI-LM, In ICO SystemController, judgment processing of
a GENIVI-LM client has an error, surface id was doubly assigned to surface,
and the client has carried out the end of an error.
Repair corrected judgment processing of the GENIVI-LM client.
Moreover, it united, and since specification processing of the buffer size
of surface had an error, it corrected.
In addition, these are the problems of SystemController of ICO and these
problems do not occur in SystemController of Murphy.

Change-Id: Ide48fe8803db832972a6581b006ea40f32eb3846
Signed-off-by: Ryota Okubi <ryota.okubi@mail.toyota-td.jp>